Add an exclude patttern for wheels
Allows files to be included in the sdist but not installed.
Useful for established projects that have tests as a subdirectory of the
package directory.
